Adds documentation to new encoder functions
Warnings are now enabled for some of the deprecated functions.
The LinSignal type has a reference to the encoding type that will be populated
by the parser using the signal representation object.
Also added a few new items to the PR checklist as version number bumps and
documentation changes were often missed.